+++ spond_classes/event.py
@@ -1,9 +1,19 @@
 """Module containing `Event` class and related `EventType`,`Responses` classes."""
 
+import sys
+
+if sys.version_info < (3, 11):
+    from typing_extensions import Self
+else:
+    from typing import Self
+
+
 from datetime import datetime
 from enum import Enum
 
 from pydantic import BaseModel, Field
+
+from .types import DictFromJSON
 
 
 class Responses(BaseModel):
@@ -11,15 +21,20 @@ class Responses(BaseModel):
 
     # Lists which always exist in API data, but may be empty
     accepted_uids: list[str] = Field(alias="acceptedIds")
-    """`acceptedIds` in API."""
+    """`acceptedIds` in Spond API.
+    May be empty."""
     declined_uids: list[str] = Field(alias="declinedIds")
-    """`declinedIds` in API."""
+    """`declinedIds` in Spond API.
+    May be empty."""
     unanswered_uids: list[str] = Field(alias="unansweredIds")
-    """`unansweredIds` in API."""
+    """`unansweredIds` in Spond API.
+    May be empty."""
     waiting_list_uids: list[str] = Field(alias="waitinglistIds")
-    """`waitinglistIds` in API."""
+    """`waitinglistIds` in Spond API.
+    May be empty."""
     unconfirmed_uids: list[str] = Field(alias="unconfirmedIds")
-    """`unconfirmedIds` in API."""
+    """`unconfirmedIds` in Spond API.
+    May be empty."""
 
 
 class EventType(Enum):
@@ -30,32 +45,29 @@ class EventType(Enum):
 
 
 class Event(BaseModel):
-    """Represents an event in the Spond system.
-
-    Events data is retrieved from the `events` API endpoint.
-    """
+    """Represents an event in the Spond system."""
 
     uid: str = Field(alias="id")
-    """`id` in API; aliased as that's a Python built-in, and the Spond package
+    """`id` in Spond API; aliased as that's a Python built-in, and the Spond package
     uses `uid`."""
     heading: str
     responses: Responses
     type: EventType
     created_time: datetime = Field(alias="createdTime")
-    """Derived from `createdTime` in API."""
+    """Derived from `createdTime` in Spond API."""
     end_time: datetime = Field(alias="endTimestamp")
     """Datetime at which the `Event` ends.
-        Derived from `endTimestamp` in API."""
+    Derived from `endTimestamp` in Spond API."""
     start_time: datetime = Field(alias="startTimestamp")
     """Datetime at which the `Event` starts.
-    Derived from `startTimestamp` in API."""
+    Derived from `startTimestamp` in Spond API."""
 
     # Optional in API data
     cancelled: bool | None = Field(default=None)
-    """Optional."""
+    """Not always present."""
     invite_time: datetime | None = Field(alias="inviteTime", default=None)
-    """Optional.
-    Derived from `inviteTime` in API."""
+    """Derived from `inviteTime` in Spond API.
+    Not always present."""
 
     def __str__(self) -> str:
         """Return simple human-readable description.
@@ -73,6 +85,22 @@ class Event(BaseModel):
     def url(self) -> str:
         """Return the URL of the `Event`, for convenience."""
         return f"https://spond.com/client/sponds/{self.uid}/"
+
+    @classmethod
+    def from_dict(cls, dict_: DictFromJSON) -> Self:
+        """Construct an `Event`.
+
+        Parameters
+        ----------
+        dict_
+            as returned by `spond.spond.Spond.get_event()`
+            or from the list returned by `spond.spond.Spond.get_events()`.
+
+        Returns
+        -------
+        `Event`
+        """
+        return cls(**dict_)
 
 
 class MatchType(Enum):
@@ -99,4 +127,3 @@ class Match(Event):
     """Represents match event."""
     match_event: bool = Field(alias="matchEvent", default=True)
     match_info: MatchInfo = Field(alias="matchInfo")
-