+++ tests/__init__.py
@@ -1,6 +1,1 @@
 """Test suite."""
-
-from typing import Any, TypeAlias
-
-DictFromJSON: TypeAlias = dict[str, Any]
-"""Simple type alias to annotate data retrieved from the API."""