+++ tests/test_member.py
@@ -2,16 +2,63 @@
 
 from datetime import datetime, timezone
 
-from spond_classes import Member
+import pytest
 
-from . import DictFromJSON
+from spond_classes import Member
+from spond_classes.types import DictFromJSON
+
+
+@pytest.fixture
+def simple_member_data() -> DictFromJSON:
+    """Simplest possible member data in this implementation.
+
+    Mocks dict returned by `Spond.get_group()['members'][n]`.
+    """
+    return {
+        "id": "M1",
+        "firstName": "Brendan",
+        "lastName": "Gleason",
+        "subGroups": [],
+        "createdTime": "2022-03-24T16:36:29Z",
+    }
+
+
+@pytest.fixture
+def complex_member_data() -> DictFromJSON:
+    """Member data with all implemented fields populated.
+
+    Mocks dict returned by `Spond.get_group()['members'][n]`.
+    """
+    return {
+        "id": "M2",
+        "firstName": "Ciarán",
+        "lastName": "Hinds",
+        "createdTime": "2022-03-24T16:36:29Z",
+        # optional:
+        "email": "ciarán@example.com",
+        "phoneNumber": "+123456789",
+        "profile": {
+            "id": "M2P2",
+            "firstName": "",
+            "lastName": "",
+            # optional:
+            "email": "ciarán2@example.com",
+            "phoneNumber": "+123456790",
+        },
+        "roles": [
+            "M2R2",
+        ],
+        "subGroups": [
+            "M2S2",
+        ],
+    }
 
 
 def test_from_dict_simple(simple_member_data: DictFromJSON) -> None:
     """Test that Member is created from the simplest possible data."""
     # arrange
     # act
-    my_member = Member.model_validate(simple_member_data)
+    my_member = Member(**simple_member_data)
     # assert
     assert my_member.uid == "M1"
     assert my_member.created_time == datetime(
@@ -35,7 +82,7 @@ def test_from_dict_full(complex_member_data: DictFromJSON) -> None:
     """Test that Member is created from dict with all supported attributes."""
     # arrange
     # act
-    my_member = Member.model_validate(complex_member_data)
+    my_member = Member(**complex_member_data)
     # assert
     assert my_member.uid == "M2"
     assert my_member.created_time == datetime(