+++ tests/test_role.py
@@ -1,15 +1,28 @@
 """Tests for Role class."""
 
-from spond_classes import Role
+import pytest
 
-from . import DictFromJSON
+from spond_classes import Role
+from spond_classes.types import DictFromJSON
+
+
+@pytest.fixture
+def simple_role_data() -> DictFromJSON:
+    """Simplest possible role data in this implementation.
+
+    Mocks dict returned by `Spond.get_group()['roles'][n]`.
+    """
+    return {
+        "id": "R1",
+        "name": "Role One",
+    }
 
 
 def test_from_dict(simple_role_data: DictFromJSON) -> None:
     """Test that Role is created from the simplest possible data dict."""
     # arrange
     # act
-    my_role = Role.model_validate(simple_role_data)
+    my_role = Role(**simple_role_data)
     # assert
     assert my_role.uid == "R1"
     assert my_role.name == "Role One"